Move game_message_type to enums.h
This reduces the extent to which output.h need be included (which is a
fairly expensive header).
